+++ packages/react/src/components/InitialScreen/src/components/InitialScreenBullet/initialScreenBullet.types.ts
@@ -1,0 +1,11 @@
+import { BoxProps } from "@nimbus-ds/components";
+import { PropsWithChildren, ReactNode } from "react";
+
+export type BulletProps = {
+  icon: ReactNode;
+  text: string;
+};
+
+export type InitialScreenBulletProps = PropsWithChildren<
+  BulletProps & Omit<BoxProps, "display" | "gap">
+>;