+++ packages/react/src/components/InitialScreen/src/components/InitialScreenFeature/initialScreenFeature.types.ts
@@ -1,0 +1,17 @@
+import { BoxProps } from "@nimbus-ds/components";
+import { PropsWithChildren, ReactNode } from "react";
+
+export type InitialScreenFeatureItemProps = PropsWithChildren<{
+  title: string;
+  description: string;
+}>;
+
+export type InitialScreenFeatureProps = {
+  content: ReactNode;
+  image: ReactNode;
+};
+
+export type InitialScreenFeatureItemSpacingProps = Omit<
+  BoxProps,
+  "borderTopWidth" | "borderBottomWidth" | "borderColor" | "borderStyle"
+>;