+++ packages/react/src/components/InitialScreen/src/components/InitialScreenHero/index.ts
@@ -1,0 +1,5 @@
+import { InitialScreenHero } from "./InitialScreenHero";
+
+export { InitialScreenHero };
+
+export default InitialScreenHero;