+++ packages/react/src/components/InitialScreen/src/components/InitialScreenSection/index.ts
@@ -1,0 +1,5 @@
+import { InitialScreenSection } from "./InitialScreenSection";
+
+export { InitialScreenSection };
+
+export default InitialScreenSection;