+++ packages/react/src/components/InitialScreen/src/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./InitialScreenHero";
+export * from "./InitialScreenBullet";
+export * from "./InitialScreenSection";
+export * from "./InitialScreenCard";
+export * from "./InitialScreenFeature";
+export * from "./InitialScreenModule";