+++ packages/react/src/components/InitialScreen/src/index.ts
@@ -1,0 +1,7 @@
+import { InitialScreen } from "./InitialScreen";
+
+export { InitialScreen } from "./InitialScreen";
+
+export type { InitialScreenProps } from "./initialScreen.types";
+
+export default InitialScreen;